refactor: step component refactor (#44)
